MGDOBR-865: Run tests in the 'mocked' folder only with EnvType.Mocked
